[#162790166] specific inheritance fix for shared_examples
We were testing for parent/child relationship	between our resources and 
FrederickAPI::V2::Resource, but background jobs introduces a new parent 
possibility with FrederickAPI::V2::Resource as the grandparent. Instead 
of ensuring the superclass be there, we are ensuring that somewhere in the
class ancestry there is FrederickAPI::V2::Resource.